logging: Allow forcing color output
`p` stands for `print color` obviously.
The engine interface looks pretty sleek so rather than passing this
flag down to the logger, just pulled in fatih/color to the main.
/shrug

TEST=manual
Ran with and without `-p` in an environment where color was previously
suppressed due to not appearing to be a TTY, and now there's color when
appropriate